% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender-data.R
\docType{data}
\name{napp}
\alias{napp}
\title{North Atlantic Population Project names dataset}
\format{A data frame with c. 650k observations and 5 variables}
\source{
Minnesota Population Center.
  \href{https://www.nappdata.org/napp/}{North Atlantic Population Project:
  Complete Count Microdata}. Version 2.0 [Machine-readable database].
  Minneapolis: Minnesota Population Center, 2008.
}
\usage{
napp
}
\description{
A dataset containing the number of instances of male and female names and
birth years drawn from census microdata from Canada, Great Britain,
Germany, Iceland, Norway, and Sweden from 1801 to 1910. This data is taken
from the \href{https://www.nappdata.org/napp/}{North Atlantic Population
Project}.
}
\keyword{datasets}

