% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender-data.R
\docType{data}
\name{ssa_state}
\alias{ssa_state}
\title{Social Security Administration state names dataset}
\format{A data frame with 5,267,234 observations and 5 variables}
\source{
Social Security Administration,
  \url{http://catalog.data.gov/dataset/baby-names-from-social-security-card-applications-data-by-state-and-district-of-}
}
\usage{
ssa_state
}
\description{
A dataset containing the number of instances of male and female names born
in each state for the years 1910 to 2012 for people who have received Social
Security Numbers. The SSA includes only names were used more than five times
in a given state in a given year.
}
\keyword{datasets}

